moved iai robot configs into iai_robots subfolder
