+++ scripts/iai_robots/pr2/pr2_mujoco_open_loop.py
@@ -3,7 +3,7 @@ import rospy
 
 from giskardpy.configs.behavior_tree_config import OpenLoopConfig
 from giskardpy.configs.giskard import Giskard
-from giskardpy.configs.pr2 import PR2CollisionAvoidance
+from giskardpy.configs.iai_robots.pr2 import PR2CollisionAvoidance
 from giskardpy.configs.qp_controller_config import QPControllerConfig
 from giskardpy.configs.world_config import WorldWithOmniDriveRobot
