+++ src/giskardpy/configs/iai_robots/boxy.py
