+++ src/giskardpy/configs/iai_robots/hsr.py
