+++ test/test_integration_donbot.py
@@ -3,20 +3,18 @@ from typing import Optional
 import numpy as np
 import pytest
 import rospy
-from geometry_msgs.msg import PoseStamped, Point, Quaternion, Pose, Vector3Stamped, PointStamped
-from std_srvs.srv import Trigger
-from tf.transformations import quaternion_from_matrix, quaternion_about_axis, quaternion_matrix, rotation_from_matrix
+from geometry_msgs.msg import PoseStamped, Point, Quaternion, Vector3Stamped, PointStamped
+from tf.transformations import quaternion_about_axis, quaternion_matrix, rotation_from_matrix
 
 import giskardpy.utils.tfwrapper as tf
-from giskard_msgs.msg import MoveResult, CollisionEntry
-from giskardpy import identifier
+from giskard_msgs.msg import MoveResult
 from giskardpy.configs.behavior_tree_config import StandAloneConfig
-from giskardpy.configs.donbot import WorldWithBoxyBaseConfig, DonbotCollisionAvoidanceConfig, DonbotStandaloneInterfaceConfig
+from giskardpy.configs.iai_robots.donbot import WorldWithBoxyBaseConfig, DonbotCollisionAvoidanceConfig, DonbotStandaloneInterfaceConfig
 from giskardpy.configs.giskard import Giskard
 from giskardpy.configs.qp_controller_config import QPControllerConfig
-from giskardpy.goals.goal import WEIGHT_BELOW_CA, WEIGHT_ABOVE_CA
+from giskardpy.goals.goal import WEIGHT_BELOW_CA
 from giskardpy.utils.utils import launch_launchfile
-from utils_for_tests import compare_poses, GiskardTestWrapper
+from utils_for_tests import GiskardTestWrapper
 
 # TODO roslaunch iai_donbot_sim ros_control_sim.launch
