+++ test/test_integration_hsr.py
@@ -3,21 +3,14 @@ from typing import Optional
 
 import numpy as np
 import pytest
-import rospy
-from geometry_msgs.msg import PoseStamped, Point, Quaternion, PointStamped, Vector3Stamped, Pose
+from geometry_msgs.msg import PoseStamped, Point, Quaternion, PointStamped, Vector3Stamped
 from numpy import pi
-from std_srvs.srv import Trigger
-from tf.transformations import quaternion_from_matrix, quaternion_about_axis, rotation_from_matrix, quaternion_matrix
+from tf.transformations import quaternion_from_matrix, quaternion_about_axis
 
-import giskardpy.utils.tfwrapper as tf
 from giskardpy.configs.behavior_tree_config import StandAloneConfig
 from giskardpy.configs.giskard import Giskard
-from giskardpy.configs.hsr import HSRCollisionAvoidance, WorldWithHSRConfig, HSRStandaloneInterface
+from giskardpy.configs.iai_robots.hsr import HSRCollisionAvoidance, WorldWithHSRConfig, HSRStandaloneInterface
 from giskardpy.configs.qp_controller_config import QPControllerConfig
-from giskardpy.configs.world_config import WorldWithOmniDriveRobot
-from giskardpy.model.utils import make_world_body_box
-from giskardpy.my_types import Derivatives
-from giskardpy.python_interface import GiskardWrapper
 from giskardpy.utils.utils import launch_launchfile
 from utils_for_tests import compare_poses, GiskardTestWrapper
