+++ test/test_integration_pr2_mujoco.py
@@ -12,7 +12,7 @@ import giskardpy.utils.tfwrapper as tf
 from giskard_msgs.msg import MoveResult, MoveGoal
 from giskardpy.configs.behavior_tree_config import OpenLoopConfig
 from giskardpy.configs.giskard import Giskard
-from giskardpy.configs.pr2 import PR2CollisionAvoidance, PR2JointTrajServerMujocoInterface
+from giskardpy.configs.iai_robots.pr2 import PR2CollisionAvoidance, PR2JointTrajServerMujocoInterface
 from giskardpy.configs.qp_controller_config import QPControllerConfig
 from giskardpy.configs.world_config import WorldWithOmniDriveRobot
 from test_integration_pr2 import PR2TestWrapper, TestJointGoals, pocky_pose
