+++ test/test_integration_pr2_mujoco_closed_loop.py
@@ -13,7 +13,7 @@ import giskardpy.utils.tfwrapper as tf
 from giskard_msgs.msg import MoveResult, MoveGoal
 from giskardpy.configs.behavior_tree_config import ClosedLoopConfig
 from giskardpy.configs.giskard import Giskard
-from giskardpy.configs.pr2 import PR2CollisionAvoidance, PR2VelocityMujocoInterface
+from giskardpy.configs.iai_robots.pr2 import PR2CollisionAvoidance, PR2VelocityMujocoInterface
 from giskardpy.configs.qp_controller_config import QPControllerConfig
 from giskardpy.configs.world_config import WorldWithOmniDriveRobot
 from giskardpy.data_types import JointStates
